+++ OArchiveImpl.h
@@ -56,7 +56,7 @@ namespace leap {
     void WriteByteArray(const void* pBuf, uint64_t ncb, bool writeSize = false) override;
     void WriteString(const void* pbuf, uint64_t charCount, uint8_t charSize) override;
     void WriteBool(bool value) override;
-    void WriteInteger(int64_t value, size_t ncb) override;
+    void WriteInteger(int64_t value, uint8_t ncb) override;
     void WriteArray(const field_serializer& desc, uint64_t n, std::function<const void*()> enumerator) override;
     void WriteDictionary(uint64_t n, const field_serializer& keyDesc, std::function<const void*()> keyEnumerator, const field_serializer& valueDesc, std::function<const void*()> valueEnumerator) override;
     
@@ -65,11 +65,11 @@ namespace leap {
     uint64_t SizeArray(const field_serializer& desc, uint64_t n, std::function<const void*()> enumerator) const override;
     uint64_t SizeDictionary(uint64_t n, const field_serializer& keyDesc, std::function<const void*()> keyEnumerator, const field_serializer& valueDesc, std::function<const void*()> valueEnumerator) const override;
     
-    inline size_t SizeString(const void* pBuf, uint64_t charCount, size_t charSize) const override { return (size_t)(sizeof(uint32_t) + (charCount*charSize)); }
-    size_t SizeInteger(int64_t value, size_t ncb) const override;
-    size_t SizeFloat(float value) const override { return sizeof(float); }
-    size_t SizeFloat(double value) const override { return sizeof(double); }
-    size_t SizeBool(bool) const override { return 1; }
+    inline uint64_t SizeString(const void* pBuf, uint64_t charCount, uint8_t charSize) const override { return (size_t)(sizeof(uint32_t) + (charCount*charSize)); }
+    uint64_t SizeInteger(int64_t value, uint8_t ncb) const override;
+    uint64_t SizeFloat(float value) const override { return sizeof(float); }
+    uint64_t SizeFloat(double value) const override { return sizeof(double); }
+    uint64_t SizeBool(bool) const override { return 1; }
 
     /// <summary>
     /// Processes objects on the internal queue until the queue is empty
